immunogenetr: A Comprehensive Toolkit for Clinical HLA Informatics

A comprehensive toolkit for clinical Human Leukocyte Antigen (HLA) informatics, built on 'tidyverse' https://tidyverse.tidyverse.org/ principles and making use of genotype list string (GL string, Mack et al. (2023) doi:10.1111/tan.15126 ) for storing and computing HLA genotype data. Specific functionalities include: coercion of HLA data in tabular format to and from GL string; calculation of matching and mismatching in all directions, with multiple output formats; automatic formatting of HLA data for searching within a GL string; truncation of molecular HLA data to a specific number of fields; and reading HLA genotypes in HML files and extracting the GL string. This library is intended for research use. Any application making use of this package in a clinical setting will need to be independently validated according to local regulations.

See also

Author

Maintainer: Nicholas Brown nicholas.brown@pennmedicine.upenn.edu (ORCID)

Authors:

Busra Coskun bcoskun@sas.upenn.edu (ORCID)